Unified test files for dereferencing a member expression.
