+++ tests/dynamic_checking/bounds/subscript_member_expr_opt.c
@@ -10,7 +10,7 @@
 // optimisation is breaking the dynamic checks.
 
 //
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t1 -DTEST_READ -DARROW_OPERATOR -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t1 -DTEST_READ -DARROW_OPERATOR -Werror -O3
 // RUN: %t1 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-1-READ
 // RUN: %t1 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-2-READ
 // RUN: %t1 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-3-READ
@@ -21,7 +21,7 @@
 // RUN: %t1 fail4 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,ARROW,FAIL-4
 // RUN: %t1 fail5 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,ARROW,FAIL-5
 //
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t2 -DTEST_WRITE -DARROW_OPERATOR -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t2 -DTEST_WRITE -DARROW_OPERATOR -Werror -O3
 // RUN: %t2 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-1-WRITE
 // RUN: %t2 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-2-WRITE
 // RUN: %t2 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-3-WRITE
@@ -32,7 +32,7 @@
 // RUN: %t2 fail4 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,ARROW,FAIL-4
 // RUN: %t2 fail5 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,ARROW,FAIL-5
 
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t3 -DTEST_INCREMENT -DARROW_OPERATOR -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t3 -DTEST_INCREMENT -DARROW_OPERATOR -Werror -O3
 // RUN: %t3 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-1-INCREMENT
 // RUN: %t3 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-2-INCREMENT
 // RUN: %t3 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-3-INCREMENT
@@ -43,7 +43,7 @@
 // RUN: %t3 fail4 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,ARROW,FAIL-4
 // RUN: %t3 fail5 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,ARROW,FAIL-5
 
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t4 -DTEST_COMPOUND_ASSIGN -DARROW_OPERATOR -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t4 -DTEST_COMPOUND_ASSIGN -DARROW_OPERATOR -Werror -O3
 // RUN: %t4 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-1-COMPOUND-ASSIGN
 // RUN: %t4 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-2-COMPOUND-ASSIGN
 // RUN: %t4 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,ARROW,PASS-3-COMPOUND-ASSIGN
@@ -61,7 +61,7 @@
 // removed and new temporary names used.
 // -----------------------------------------------------------------------------------
 
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t5 -DTEST_READ -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t5 -DTEST_READ -Werror -O3
 // RUN: %t5 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-1-READ
 // RUN: %t5 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-2-READ
 // RUN: %t5 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-3-READ
@@ -72,7 +72,7 @@
 // RUN: %t5 fail4 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,FAIL-4
 // RUN: %t5 fail5 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,FAIL-5
 //
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t6 -DTEST_WRITE -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t6 -DTEST_WRITE -Werror -O3
 // RUN: %t6 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-1-WRITE
 // RUN: %t6 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-2-WRITE
 // RUN: %t6 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-3-WRITE
@@ -83,7 +83,7 @@
 // RUN: %t6 fail4 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,FAIL-4
 // RUN: %t6 fail5 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,FAIL-5
 
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t7 -DTEST_INCREMENT -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t7 -DTEST_INCREMENT -Werror  -O3
 // RUN: %t7 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-1-INCREMENT
 // RUN: %t7 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-2-INCREMENT
 // RUN: %t7 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-3-INCREMENT
@@ -94,7 +94,7 @@
 // RUN: %t7 fail4 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,FAIL-4
 // RUN: %t7 fail5 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-FAIL,FAIL-5
 
-// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t8 -DTEST_COMPOUND_ASSIGN -Werror
+// RUN: %clang -fcheckedc-extension %S/subscript_member_expr.c -o %t8 -DTEST_COMPOUND_ASSIGN -Werror -O3
 // RUN: %t8 pass1 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-1-COMPOUND-ASSIGN
 // RUN: %t8 pass2 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-2-COMPOUND-ASSIGN
 // RUN: %t8 pass3 | FileCheck %S/subscript_member_expr.c --check-prefixes=CHECK,CHECK-PASS,PASS-3-COMPOUND-ASSIGN
